fix: don't let uglify-es inline functions with arguments
Fixes #346.

This is a workaround for a bug in uglifyjs2
(https://github.com/mishoo/UglifyJS2/issues/2842), which can cause name
collisions when a function with arguments is inlined. This can cause an
unintended shadowing of a `var` or `let`, or a `TypeError: Assignment to
constant variable` in case of a `const`.